[SPARK-5376][Mesos] MesosExecutor should have correct resources - Divided task and executor resources - Added `spark.mesos.executor.cpus` and fixed docs

+++ core/src/test/scala/org/apache/spark/scheduler/mesos/MesosSchedulerBackendSuite.scala
@@ -99,7 +99,7 @@ class MesosSchedulerBackendSuite extends FunSuite with LocalSparkContext with Ea
     EasyMock.expect(sc.listenerBus).andReturn(listenerBus)
     EasyMock.replay(sc)
 
-    val minMem = MemoryUtils.calculateTotalMemory(sc).toInt
+    val minMem = MemoryUtils.calculateTotalMemory(sc)
     val minCpu = 4
 
     val mesosOffers = new java.util.ArrayList[Offer]
